+++ src/components/Comments/Comments.scss
@@ -4,11 +4,9 @@
 .comments {
   margin-bottom: 4px;
   font-size: 14px;
-  font-family: 'Noto Sans KR', sans-serif;
 
   .details {
     margin-left: 3px;
     line-height: 120%;
-    font-family: 'Noto Sans KR', sans-serif;
   }
 }
